Добавила компонент с табами, отрисовку карточки фильма по id
